rename "repo" to "pack"
