+++ packs/anilist/graphql.ts
