+++ packs/anilist/index.ts
