+++ packs/anilist/lastPage.ts
