+++ packs/anilist/types.ts
