+++ tests/packs.test.ts
@@ -2,4 +2,4 @@
 
 // import { stub } from 'https://deno.land/std@0.168.0/testing/mock.ts';
 
-// import repo from '../src/repo.ts';
+// import packs from '../src/packs.ts';
